Password.java
The project is about password Hashing i.e. hashing the password. 
If a user created a new account then the password he enters will be hashed i.e. converted into characters which cannot be retrieved. 
Use - if there is a leakage of database or is attacked by hackers. Then the hackers can only access the hashed password which is irreversible i.e. hacker cannot obtain the original password from the original password.